Move tests to single file
